refactor: move the remainder of `shared` into `stdlib`
