+++ src/lib.rs
@@ -11,7 +11,6 @@ mod completion;
 mod diagnostics;
 mod lsp;
 mod server;
-mod shared;
 mod stdlib;
 mod visitors;
 #[cfg(feature = "wasm")]
